+++ locksmith/src/fulfillment/dispatcher.ts
@@ -94,7 +94,7 @@ export default class Dispatcher {
           gasPrice.mul(GAS_COST)
         )}) on ${network}`
       )
-      return false
+      return true
     }
     return true
   }
